[Refactor] directory 구조 재세팅 #177
